+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.1.0] - TBD
+### Added
+- Logical folding for maven log in debug mode
 ### Changed
 - Improved Downloading/Progress/Downloaded log lines behavior
 ### Fixed